+++ backend/src/main/java/codesquad/team4/issuetracker/exception/GlobalExceptionHandler.java
@@ -3,6 +3,7 @@ package codesquad.team4.issuetracker.exception;
 import codesquad.team4.issuetracker.exception.badrequest.InvalidRequestException;
 import codesquad.team4.issuetracker.exception.conflict.ConflictException;
 import codesquad.team4.issuetracker.exception.notfound.DataNotFoundException;
+import codesquad.team4.issuetracker.exception.unauthorized.UnauthorizedException;
 import codesquad.team4.issuetracker.response.ApiResponse;
 import io.swagger.v3.oas.annotations.Hidden;
 import org.springframework.http.HttpStatus;
@@ -30,6 +31,12 @@ public class GlobalExceptionHandler {
     @ExceptionHandler(ConflictException.class)
     @ResponseStatus(HttpStatus.CONFLICT)
     public ApiResponse<String> handleConflictException(ConflictException ex) {
+        return ApiResponse.fail(ex.getMessage());
+    }
+
+    @ExceptionHandler(UnauthorizedException.class)
+    @ResponseStatus(HttpStatus.UNAUTHORIZED)
+    public ApiResponse<String> handleUnauthorizedException(UnauthorizedException ex) {
         return ApiResponse.fail(ex.getMessage());
     }
 }
