+++ backend/src/main/java/codesquad/team4/issuetracker/exception/unauthorized/PasswordNotEqualException.java
@@ -1,8 +1,8 @@
-package codesquad.team4.issuetracker.exception.badrequest;
+package codesquad.team4.issuetracker.exception.unauthorized;
 
 import static codesquad.team4.issuetracker.exception.ExceptionMessage.USER_PASSWORD_NOT_EQUAL;
 
-public class PasswordNotEqualException extends InvalidRequestException {
+public class PasswordNotEqualException extends UnauthorizedException {
     public PasswordNotEqualException() {
         super(USER_PASSWORD_NOT_EQUAL);
     }
