+++ backend/src/main/java/codesquad/team4/issuetracker/exception/unauthorized/UnauthorizedException.java
@@ -1,0 +1,7 @@
+package codesquad.team4.issuetracker.exception.unauthorized;
+
+abstract public class UnauthorizedException extends RuntimeException{
+    public UnauthorizedException(String message) {
+        super(message);
+    }
+}
